Extract expression processor.
